gcloud/functions/metoffice/requirements.txt
requests
datetime
numpy
Pillow
google-cloud-storage